[fix] Close #43
